DOODLEROUNDHOUSE
its not a cinematic thing like anthony envisioned but i really can't see grabs being a good thing in this (plus itll be real annoying to feel with the latency) so a knockback that functions is pretty good me thinks

+++ Assets/scripts/MainGameScript/BuffDebuffSystem/BuffBase.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [System.Serializable]
 public abstract class BuffBase : ScriptableObject
 {
-    public int duration;
+    public float duration;
     public bool combatOnly = false;
     public bool stackable = false;
     public GameObject buffFx;
@@ -87,5 +87,5 @@ public abstract class BuffBase : ScriptableObject
 
     public virtual void OnEveryTick(AbilityManager stats) { }
 
-    public virtual int GetDuration(EntityStats stats) { return duration; }
+    public virtual float GetDuration(EntityStats stats) { return duration; }
 }
